[Refactor] AddRecord 관련 코드 작성 with TCA
- 현재 네비게이션 연결 안됨
- photopicker, pickermap 부분 코드 TCA 에 맞춰 수정
( pickermap 내부에 일부분 TCA 아닌 기존 코드 사용 )
