+++ Moment/Moment/AddRecord/PhotoPicker.swift
@@ -8,14 +8,12 @@
 import SwiftUI
 import PhotosUI
 
+import ComposableArchitecture
+
 // MARK: - PHPickerViewController 로, 앨범에서 사진 고르기 사용
 struct PhotoPicker: UIViewControllerRepresentable {
-    @Binding var selectedPhotos: [UIImage?]
-    @Binding var isLibraryPresented: Bool
-    private let tatoalSpace = 3
-    let remainingSpaces: Int
-    
-
+    @Bindable var store: StoreOf<AddRecordViewFeature>
+        
     class Coordinator: PHPickerViewControllerDelegate {
         private let parent: PhotoPicker
         
@@ -30,12 +28,13 @@ struct PhotoPicker: UIViewControllerRepresentable {
                 _ = itemProvider.loadObject(ofClass: UIImage.self) { (image, error) in
                     if let image = image as? UIImage {
                         DispatchQueue.main.async {
-                            self.parent.selectedPhotos[self.parent.tatoalSpace - self.parent.remainingSpaces + idx] = image
+                            self.parent.store.selectedImages.append(image)
+//                            self.parent.selectedPhotos[self.parent.tatoalSpace - self.parent.remainingSpaces + idx] = image
                         }
                     }
                 }
             }
-            parent.isLibraryPresented = false
+            parent.store.isPhotoPickerSheet = false
         }
     }
     
@@ -46,7 +45,7 @@ struct PhotoPicker: UIViewControllerRepresentable {
     func makeUIViewController(context: Context) -> PHPickerViewController {
         var configuration = PHPickerConfiguration(photoLibrary: .shared())
         configuration.filter = PHPickerFilter.any(of: [.images])
-        configuration.selectionLimit = remainingSpaces
+        configuration.selectionLimit = store.maxSelectImageCount
         configuration.preferredAssetRepresentationMode = .current
         let controller = PHPickerViewController(configuration: configuration)
         controller.delegate = context.coordinator
